backend/requirements.txt
# Vision Agents SDK + Plugins (PRIMARY intelligence layer)
vision-agents[getstream,gemini,deepgram,elevenlabs,ultralytics]

# Web server + WebSocket
fastapi[standard]
uvicorn[standard]
websockets

# Environment
python-dotenv

# CV (fallback only — SDK is primary)
opencv-python-headless
numpy
mediapipe

# AV for av.VideoFrame (required by SDK video pipeline)
av

# WebRTC (required by SDK edge transport)
aiortc
